remove core library, configs not reusable
