--- emis-client-core/spotbugs-excludes.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<FindBugsFilter>
-  <Match>
-    <Class name="gov.va.api.lighthouse.emis.EmisClient"/>
-    <Bug pattern="URLCONNECTION_SSRF_FD"/>
-  </Match>
-</FindBugsFilter>
