--- emis-client-core/src/main/java/gov/va/api/lighthouse/emis/EmisClient.java
@@ -1,83 +1,0 @@
-package gov.va.api.lighthouse.emis;
-
-import java.io.InputStream;
-import java.net.Socket;
-import java.security.KeyStore;
-import java.security.Principal;
-import java.security.PrivateKey;
-import java.security.SecureRandom;
-import java.security.cert.X509Certificate;
-import javax.net.ssl.KeyManager;
-import javax.net.ssl.KeyManagerFactory;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManagerFactory;
-import javax.net.ssl.X509KeyManager;
-import lombok.SneakyThrows;
-import org.springframework.util.ResourceUtils;
-
-public class EmisClient {
-  /**
-   * Utility method to create SSL Context to be used by a client. @Param EmisConfig from client
-   * properties.
-   */
-  @SneakyThrows
-  public static SSLContext createSslContext(EmisConfig config) {
-    try (InputStream keystoreInputStream =
-            ResourceUtils.getURL(config.getKeystorePath()).openStream();
-        InputStream truststoreInputStream =
-            ResourceUtils.getURL(config.getTruststorePath()).openStream()) {
-      // Keystore
-      KeyStore keystore = KeyStore.getInstance("JKS");
-      keystore.load(keystoreInputStream, config.getKeystorePassword().toCharArray());
-      KeyManagerFactory kmf =
-          KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
-      kmf.init(keystore, config.getKeystorePassword().toCharArray());
-      final X509KeyManager origKm = (X509KeyManager) kmf.getKeyManagers()[0];
-      // Truststore
-      KeyStore truststore = KeyStore.getInstance("JKS");
-      truststore.load(truststoreInputStream, config.getTruststorePassword().toCharArray());
-      TrustManagerFactory trustManagerFactory =
-          TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
-      trustManagerFactory.init(truststore);
-      // SSL Context
-      SSLContext sslContext = SSLContext.getInstance("TLS");
-      X509KeyManager keyManager =
-          new X509KeyManager() {
-            @Override
-            public String chooseClientAlias(String[] keyType, Principal[] issuers, Socket socket) {
-              return config.getKeyAlias();
-            }
-
-            @Override
-            public String chooseServerAlias(String keyType, Principal[] issuers, Socket socket) {
-              return origKm.chooseServerAlias(keyType, issuers, socket);
-            }
-
-            @Override
-            public X509Certificate[] getCertificateChain(String alias) {
-              return origKm.getCertificateChain(alias);
-            }
-
-            @Override
-            public String[] getClientAliases(String keyType, Principal[] issuers) {
-              return origKm.getClientAliases(keyType, issuers);
-            }
-
-            @Override
-            public PrivateKey getPrivateKey(String alias) {
-              return origKm.getPrivateKey(alias);
-            }
-
-            @Override
-            public String[] getServerAliases(String keyType, Principal[] issuers) {
-              return origKm.getServerAliases(keyType, issuers);
-            }
-          };
-      sslContext.init(
-          new KeyManager[] {keyManager},
-          trustManagerFactory.getTrustManagers(),
-          new SecureRandom());
-      return sslContext;
-    }
-  }
-}
