+++ emis-client-v1/spotbugs-excludes.xml
@@ -1,2 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<FindBugsFilter></FindBugsFilter>
+<FindBugsFilter>
+  <Match>
+    <Class name="gov.va.api.lighthouse.emis.SoapEmisVeteranStatusServiceClient"/>
+    <Bug pattern="URLCONNECTION_SSRF_FD"/>
+  </Match>
+</FindBugsFilter>
