+++ emis-client-v1/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  gov.va.api.lighthouse.emis.EmisConfigV1
