+++ emis-client-v2/pom.xml
@@ -13,7 +13,7 @@
   <dependencies>
     <dependency>
       <groupId>gov.va.api.lighthouse</groupId>
-      <artifactId>emis-client-core</artifactId>
+      <artifactId>emis-model</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
