+++ emis-client-v2/src/main/java/gov/va/api/lighthouse/emis/EmisConfigV2.java
@@ -1,0 +1,28 @@
+package gov.va.api.lighthouse.emis;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.Accessors;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableConfigurationProperties
+@ConfigurationProperties("emis.v2")
+@Data
+@Accessors(fluent = false)
+@NoArgsConstructor
+@Builder
+@AllArgsConstructor
+public class EmisConfigV2 {
+  String url;
+  String wsdlLocation;
+  String keystorePath;
+  String keystorePassword;
+  String keyAlias;
+  String truststorePath;
+  String truststorePassword;
+}
