+++ emis-client-v2/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  gov.va.api.lighthouse.emis.EmisConfig
+  gov.va.api.lighthouse.emis.EmisConfigV2
