+++ pom.xml
@@ -12,7 +12,6 @@
   <packaging>pom</packaging>
   <modules>
     <module>emis-model</module>
-    <module>emis-client-core</module>
     <module>emis-client-v1</module>
     <module>emis-client-v2</module>
   </modules>
